test(acp): cover fleet registry and minimal launchers
X-Agent: codex/6130-acp-routine
